requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.2
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
autogen==0.3.0
autopep8==2.3.1
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.1.3
bidict==0.23.1
blinker==1.6.2
Brotli==1.0.9
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
deepgram-sdk==3.1.4
Deprecated==1.2.13
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.0
executing==2.1.0
fastapi==0.103.0
filelock==3.13.1
FLAML==2.1.2
Flask==3.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
flatbuffers==2.0
fonttools==4.53.0
freetype-py==2.3.0
frozenlist==1.4.0
fsspec==2024.3.1
future==0.18.3
gmpy2==2.1.2
google-auth==2.29.0
googleapis-common-protos==1.56.4
graphrag==0.0.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
importlib-metadata==7.0.1
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipython==8.18.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.2.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-huggingface==0.0.3
langchain-openai==0.1.11
langchain-text-splitters==0.2.2
langsmith==0.1.82
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.0
mkl==2021.4.0
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.5
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
networkx==3.2.1
nh3==0.2.17
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.35.7
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.5
overrides==7.4.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
phi==0.6.7
pillow==10.3.0
pkginfo==1.10.0
posthog==3.5.0
priority==2.0.0
prompt_toolkit==3.0.48
protobuf==4.25.3
pulsar-client==3.5.0
pure_eval==0.2.3
py3Dmol==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyautogen==0.2.31
pycairo==1.26.1
pycodestyle==2.12.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
pypdf==5.0.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
Quart==0.19.6
quart-cors==0.7.0
rdkit==2024.3.3
rdkit-pypi==2022.9.5
readme_renderer==43.0
regex==2023.10.3
replicate==0.31.0
reportlab==4.2.2
requests==2.32.2
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
rlPyCairo==0.2.0
rsa==4.7.2
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.1.99
shellingham==1.5.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.27.0
sympy==1.12.1
taskgroup==0.0.0a4
tavily-python==0.3.3
tbb==2021.13.0
tenacity==8.4.2
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
twine==5.1.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unicodedata2==15.1.0
urllib3==1.26.19
uvicorn==0.20.0
verboselogs==1.7
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
win-inet-pton==1.1.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.3
zipp==3.19.2